EA-163 Fix test cancellazione immagini, permessi cancellazione

+++ app/Http/Controllers/ImageController.php
@@ -4,13 +4,16 @@ namespace App\Http\Controllers;
 
 use App\Models\Image;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class ImageController extends Controller
 {
     public function destroy (Image $image)
     {
-        //Controllo permessi
+        if (Auth::id() !== $image->event->author_id) {
+            abort(401);
+        }
         Storage::delete('/public/images/' . $image->file_name);
         $image->delete();
     }
